✨ (events.ts): Refactor event handling to use separate modules for mention response functionality, fetching text from URLs, and generating summary messages. This improves code organization and maintainability.
